+++ src/main/java/com/tencentcloudapi/soe/v20180724/models/SentenceInfo.java
@@ -27,7 +27,7 @@ public class SentenceInfo extends AbstractModel{
     */
     @SerializedName("SentenceId")
     @Expose
-    private Long SentenceId;
+    private Integer SentenceId;
 
     /**
     * 详细发音评估结果
@@ -65,10 +65,10 @@ public class SentenceInfo extends AbstractModel{
     private Float SuggestedScore;
 
     /**
-     * Get 句子序号，在段落、自由说模式下有效，表示断句序号，最后的综合结果的为-1. 
+     * Get 句子序号，在段落、自由说模式下有效，表示断句序号，最后的综合结果的为-1.
      * @return SentenceId 句子序号，在段落、自由说模式下有效，表示断句序号，最后的综合结果的为-1.
      */
-    public Long getSentenceId() {
+    public Integer getSentenceId() {
         return this.SentenceId;
     }
 
@@ -76,7 +76,7 @@ public class SentenceInfo extends AbstractModel{
      * Set 句子序号，在段落、自由说模式下有效，表示断句序号，最后的综合结果的为-1.
      * @param SentenceId 句子序号，在段落、自由说模式下有效，表示断句序号，最后的综合结果的为-1.
      */
-    public void setSentenceId(Long SentenceId) {
+    public void setSentenceId(Integer SentenceId) {
         this.SentenceId = SentenceId;
     }
 
@@ -169,7 +169,7 @@ public class SentenceInfo extends AbstractModel{
      */
     public SentenceInfo(SentenceInfo source) {
         if (source.SentenceId != null) {
-            this.SentenceId = new Long(source.SentenceId);
+            this.SentenceId = new Integer(source.SentenceId);
         }
         if (source.Words != null) {
             this.Words = new WordRsp[source.Words.length];
